CLOUDP-287249: IPA-123: Validate enums must be UPPER_SNAKE_CASE
