+++ tools/spectral/ipa/rulesets/functions/eachEnumValueMustBeUpperSnakeCase.js
@@ -1,23 +1,16 @@
 import { hasException } from './utils/exceptions.js';
+import {getSchemaPath, resolveObject} from './utils/componentUtils.js'
 import { casing } from '@stoplight/spectral-functions';
 
 const RULE_NAME = 'xgen-IPA-123-enum-values-must-be-upper-snake-case';
 const ERROR_MESSAGE = 'enum value must be UPPER_SNAKE_CASE.';
 
-function getSchemaPath(path) {
-  if (path.includes('components')) {
-    return path.slice(0, 3);
-  } else if (path.includes('paths')) {
-    const index = path.findIndex((item) => item === 'schema');
-    return path.slice(0, index + 1);
-  }
-}
+export default (input, _, { path, documentInventory }) => {
 
-export default (input, _, { path }) => {
-  //There are two path types for enum definition
-  //First type: components.schemas.schemaName.*.enum
-  //Second type: paths.*.method.parameters[*].schema.enum
-  if (hasException(getSchemaPath(path), RULE_NAME)) {
+  const oas = documentInventory.resolved;
+  const schemaPath = getSchemaPath(path);
+  const schemaObject = resolveObject(oas, schemaPath);
+  if (hasException(schemaObject, RULE_NAME)) {
     return;
   }
