--- tools/spectral/ipa/rulesets/functions/utils/pathUtils.js
@@ -1,14 +1,0 @@
-/**
- * Checks if a string belongs to a path parameter or a path parameter with a custom method.
- *
- * A path parameter has the format: `{paramName}`
- * A path parameter with a custom method has the format: `{paramName}:customMethod`
- *
- * @param {string} str - A string extracted from a path split by slashes.
- * @returns {boolean} True if the string matches the expected formats, false otherwise.
- */
-export function isPathParam(str) {
-  const pathParamRegEx = new RegExp(`^{[a-z][a-zA-Z0-9]*}$`);
-  const pathParamWithCustomMethodRegEx = new RegExp(`^{[a-z][a-zA-Z0-9]*}:[a-z][a-zA-Z0-9]*$`);
-  return pathParamRegEx.test(str) || pathParamWithCustomMethodRegEx.test(str);
-}
